pyproject.toml
[tool.poetry]
name = "py-connectwidgets"
version = "0.1.0"
description = ""
authors = ["Gagandeep Singh <singh.gagandeep491@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.1"
rsconnect-python = "^1.22.0"
requests = "^2.31.0"
itables = "^2.0.0"
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
